src/test/java/jpabook/jpashop/MemberRepositoryTest.java
package jpabook.jpashop;

import jpabook.jpashop.domain.Member;
import org.assertj.core.api.Assertions;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.transaction.annotation.Transactional;

@RunWith(SpringRunner.class)            // Spring 관련 실행인걸 알리기위해
@SpringBootTest                         //SpringBott test기 떄문에
public class MemberRepositoryTest {

    @Autowired MemberRepository memberRepository;

    @Test
    @Transactional                          //test 어노테이션에 같이 있으면 마지막에 rollback 처리함
    @Rollback(value = false)                //test 어노테이션에 같이 있으면 마지막에 rollback 처리하는 것을 막기위해 추가
    public void testMember() throws Exception   {
        //given
        Member member = new Member();
        member.setUsername("memberA");

        //when
        Long saveId = memberRepository.save(member);
        Member findMember = memberRepository.find(saveId);

        //then
        Assertions.assertThat(findMember.getId()).isEqualTo(member.getId());
        Assertions.assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
        Assertions.assertThat(findMember).isEqualTo(member);
        System.out.println("findMember == member :: " + (findMember == member) );
    }

}